Refactor(web-react): Extended hooks #DS-275
